+++ django_admin_filters/__init__.py
@@ -1,4 +1,4 @@
 """Embedded filters for Django Admin site."""
 from .daterange import DateRange  # noqa: F401
 from .daterange_picker import DateRangePicker  # noqa: F401
-from .multi_choice import MultiChoice  # noqa: F401
+from .multi_choice import Filter as MultiChoice  # noqa: F401
