+++ django_admin_filters/multi_choice.py
@@ -1,10 +1,11 @@
 """Django admin multi choice filter with checkboxes for db fields with choices option."""
-from django.contrib import admin
-from .mixin import Collapsed
+from .base import Filter as BaseFilter
 
 
-class MultiChoice(admin.FieldListFilter, Collapsed):
-    """Multiselect options filter.
+class Filter(BaseFilter):
+    """Multi choice options filter.
+
+    For CharField fields with 'choices' option.
 
     https://stackoverflow.com/questions/39790087/is-multi-choice-django-admin-filters-possible
     https://stackoverflow.com/questions/38508672/django-admin-filter-multiple-select
@@ -13,30 +14,20 @@ class MultiChoice(admin.FieldListFilter, Collapsed):
     """
 
     template = 'multi_choice.html'
-    parameter_name_mask = 'choice_'
+    parameter_name_mask = 'mchoice_'
 
     FILTER_LABEL = "By choices"
-    BUTTON_LABEL = "Apply"
     CHOICES_SEPARATOR = ','
 
     def __init__(self, field, request, params, model, model_admin, field_path):
-        """Customize FieldListFilter functionality."""
-        self.parameter_name = self.parameter_name_mask + field_path
+        """Extend base functionality."""
         super().__init__(field, request, params, model, model_admin, field_path)
 
-        self.title = {
-          'parameter_name': self.parameter_name,
-          'filter_name': self.FILTER_LABEL,
-          'button_label': self.BUTTON_LABEL,
-          'collapsed': self.collapsed_state,
+        self.title.update({
           'choices_separator': self.CHOICES_SEPARATOR,
-        }
-        val = self.used_parameters.get(self.parameter_name)
+        })
+        val = self.value()
         self.selected = val.split(self.CHOICES_SEPARATOR) if val else []
-
-    def expected_parameters(self):
-        """Parameter list for chice filter."""
-        return [self.parameter_name]
 
     def choices(self, changelist):
         """Define filter checkboxes."""
